requirements.txt
# Core dependencies
streamlit==1.40.0

# LangChain and related
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7
langchain-chroma==0.1.4
langchain-qdrant==0.1.4
langchain-core==0.1.53
langgraph==0.0.32
langchainhub==0.1.21

# Vector stores
chromadb==0.4.24
qdrant-client==1.7.3

# OpenAI
openai==1.12.0

# Web search
tavily-python==0.3.3

# Utilities
pydantic==2.6.1
python-dotenv==1.0.1
lark==1.1.9

# Required for parsing
tiktoken==0.5.2
